requirements.txt
flask
python-dotenv
anthropic
requests